[DSCP-92] Add index to leaf nodes
